# api/swagger.yaml
swagger: '2.0'
info:
  description: This is a server for WKON
  version: 1.0.0
  title: WKON
host: 'localhost:5088'
...
securityDefinitions:
  key:
    type: apiKey
    in: header
    name: Key
schemes:
  - http
paths:
...
  /user/createWithArray:
    post:
      security:
        - key: []
      tags:
        - user
      summary: Creates list of users with given input array
      description: ''
      operationId: createUsersWithArrayInput
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: List of user object
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: User
  /user/createWithList:
    post:
      security:
        - key: []
      tags:
        - user
      summary: Creates list of users with given input array
      description: ''
      operationId: createUsersWithListInput
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: List of user object
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: User
...
  /user/logout:
    get:
      security:
        - key: []
      tags:
        - user
      summary: Logs out current logged fin user session
      description: ''
      operationId: logoutUser
      produces:
        - application/xml
        - application/json
      parameters: []
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: User
...
  /client/login:
    get:
      security:
        - key: []
      tags:
        - client
      summary: Logs client into the system
      description: ''
      operationId: loginclient
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: true
          type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: string
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
        '400':
          description: Invalid username/password supplied
      x-swagger-router-controller: Client
...
  /project:
    post:
      security:
        - key: []
      tags:
        - project
      summary: Create project
      description: This can only be done by the logged in user.
      operationId: CreateProject
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created project object
          required: true
          schema:
            $ref: '#/definitions/Project'
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: Project
  /project/projectlist:
    get:
      security:
        - key: []
      tags:
        - project
      summary: Gets the project list which user is a member of
      description: ''
      operationId: getProjectList
      produces:
        - application/xml
        - application/json
      parameters:
        - name: userId
          in: query
          description: The user id of logged in user
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: string
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
      x-swagger-router-controller: Project
  /project/addmember:
    put:
      security:
        - key: []
      tags:
        - project
      summary: Updated project
      description: This can only be done by the logged in user.
      operationId: updateProjectByMembers
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/addmember'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: project not found
      x-swagger-router-controller: Project
  /project/deletemember:
    put:
      security:
        - key: []
      tags:
        - project
      summary: Updated project
      description: This can only be done by the logged in user.
      operationId: deleteProjectByMemberId
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: userId
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/Deletemember'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: project not found
      x-swagger-router-controller: Project
  '/project/{id}':
    get:
      security:
        - key: []
      tags:
        - project
      summary: Get project by id
      description: ''
      operationId: getProjectById
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Project'
        '400':
          description: Invalid Project Id supplied
        '404':
          description: Project not found
      x-swagger-router-controller: Project
    put:
      security:
        - key: []
      tags:
        - project
      summary: Updated project
      description: This can only be done by the logged in user.
      operationId: updateProject
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: project not found
      x-swagger-router-controller: Project
    delete:
      security:
        - key: []
      tags:
        - project
      summary: Delete project
      description: This can only be done by the logged in user.
      operationId: deleteProject
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: The project that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Invalid project supplied
        '404':
          description: Project not found
      x-swagger-router-controller: Project
  /project/getbymemberId:
    get:
      security:
        - key: []
      tags:
        - project
      summary: Get project by id
      description: ''
      operationId: getProjectByMembers
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Project'
        '400':
          description: Invalid Project Id supplied
        '404':
          description: Project not found
      x-swagger-router-controller: Project
...
  /submodule:
    post:
      security:
        - key: []
      tags:
        - submodule
      summary: submodule
      description: This enables a user to create.
      operationId: createsubmodule
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created sub module object
          required: true
          schema:
            $ref: '#/definitions/Submodule'
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: Submodule
  '/submodule/{id}':
    get:
      security:
        - key: []
      tags:
        - submodule
      summary: Get project by id
      description: ''
      operationId: getsubmodule
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Submodule'
        '400':
          description: Invalid Module Id supplied
        '404':
          description: Module not found
      x-swagger-router-controller: Submodule
    put:
      security:
        - key: []
      tags:
        - submodule
      summary: Updated module
      description: This can only be done by the logged in user.
      operationId: updatesubmodule
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/Submodule'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: project not found
      x-swagger-router-controller: Submodule
    delete:
      security:
        - key: []
      tags:
        - submodule
      summary: Delete module
      description: This can only be done by the logged in user.
      operationId: deletesubmodule
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: The module that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Invalid module supplied
        '404':
          description: module not found
      x-swagger-router-controller: Submodule
  /submodule/getbymoduleid:
    get:
      security:
        - key: []
      tags:
        - submodule
      summary: Get project by id
      description: ''
      operationId: getsubmoduleByModule
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Submodule'
        '400':
          description: Invalid Module Id supplied
        '404':
          description: Module not found
      x-swagger-router-controller: Submodule
...
definitions:
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      tags:
        type: array
        items:
          type: string
      manager:
        type: string
      role:
        type: string
    example:
      password: password
      role: role
      phone: phone
      email: email
      tags: tags
      name: username
      manager: manager
    xml:
      name: User
...
  Project:
    type: object
    properties:
      name:
        type: string
      requirement:
        type: string
      status:
        type: string
      technology:
        type: string
      expectedStartDate:
        type: string
      actualStartDate:
        type: string
      expectedEndDate:
        type: string
      actualEndDate:
        type: string
      client:
        type: string
      userId:
        type: string
    example:
      requirement: requirement
      status: status
      technology: technology
      expectedStartDate: expectedStartDate
      actualStartDate: actualStartDate
      expectedEndDate: expectedEndDate
      actualEndDate: actualEndDate
      name: name
      client: client id
      userId: user id
    xml:
      name: Project
...
  addmember:
    type: array
    items:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
    example:
      userId: userId
      role: role
    xml:
      name: addmember
...
  Deletemember:
    type: object
    properties:
      userId:
        type: string
    example:
      userId: userId
    xml:
      name: Deletemember
...
  Submodule:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      moduleId:
        type: string
    example:
      name: name
      description: description
      moduleId: moduleId
    xml:
      name: Submodule
...